split AlitheiaCoreModule into several modules
this is more in the spirit of Guice
