+++ alitheia/core/src/main/java/eu/sqooss/impl/service/admin/AdminServiceModule.java
@@ -1,0 +1,14 @@
+package eu.sqooss.impl.service.admin;
+
+import com.google.inject.AbstractModule;
+
+import eu.sqooss.service.admin.AdminService;
+
+public class AdminServiceModule extends AbstractModule {
+
+    @Override
+    protected void configure() {
+        bind(AdminService.class).to(AdminServiceImpl.class);
+    }
+    
+}
