+++ alitheia/core/src/main/java/eu/sqooss/impl/service/cluster/ClusterNodeModule.java
@@ -1,0 +1,14 @@
+package eu.sqooss.impl.service.cluster;
+
+import com.google.inject.AbstractModule;
+
+import eu.sqooss.service.cluster.ClusterNodeService;
+
+public class ClusterNodeModule extends AbstractModule {
+
+    @Override
+    protected void configure() {
+        bind(ClusterNodeService.class).to(ClusterNodeServiceImpl.class);
+    }
+
+}
