+++ alitheia/core/src/main/java/eu/sqooss/impl/service/rest/RestServiceModule.java
@@ -5,10 +5,13 @@ import org.jboss.resteasy.plugins.server.servlet.HttpServletDispatcher;
 import com.google.inject.AbstractModule;
 import com.google.inject.assistedinject.FactoryModuleBuilder;
 
+import eu.sqooss.service.rest.RestService;
+
 public class RestServiceModule extends AbstractModule {
 
 	@Override
 	protected void configure() {
+	    bind(RestService.class).to(ResteasyServiceImpl.class);
 		install(new FactoryModuleBuilder().implement(HttpServletDispatcher.class,
 				ResteasyServlet.class).build(ResteasyServletFactory.class));
 	}
