+++ alitheia/core/src/main/java/eu/sqooss/impl/service/tds/TDSServiceModule.java
@@ -1,0 +1,14 @@
+package eu.sqooss.impl.service.tds;
+
+import com.google.inject.AbstractModule;
+
+import eu.sqooss.service.tds.TDSService;
+
+public class TDSServiceModule extends AbstractModule {
+
+    @Override
+    protected void configure() {
+        bind(TDSService.class).to(TDSServiceImpl.class);
+    }
+
+}
