+++ alitheia/core/src/main/java/eu/sqooss/impl/service/updater/UpdaterServiceModule.java
@@ -5,11 +5,13 @@ import com.google.inject.assistedinject.FactoryModuleBuilder;
 
 import eu.sqooss.service.scheduler.Job;
 import eu.sqooss.service.updater.MetadataUpdater;
+import eu.sqooss.service.updater.UpdaterService;
 
-public class UpdaterModule extends AbstractModule {
+public class UpdaterServiceModule extends AbstractModule {
 
     @Override
     protected void configure() {
+        bind(UpdaterService.class).to(UpdaterServiceImpl.class);
         install(new FactoryModuleBuilder().implement(Job.class,
                 UpdaterJob.class).build(UpdaterJobFactory.class));
     }
